bot-reserva/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs>=4.15.0
aiohttp>=3.8.0
